refactor utilities for unit testing
