+++ src/main/java/com/react/pnld/services/FileService.java
@@ -30,7 +30,7 @@ public class FileService {
     private String FILE_PATH;
 
     @Autowired
-    private CSVHeadersProperties csvHeadersProperties;
+    private FileUtilService fileUtilService;
 
     @Autowired
     private FileRepository fileRepository;
@@ -57,10 +57,7 @@ public class FileService {
     }
 
     public boolean isValidCsvHeader(CsvFile csvFile) {
-
         logger.info("isValidCsvHeader. csvFile={}", csvFile);
-
-        List<String> csvFileHeaders = new ArrayList<>();
         String firstLine = "";
         try {
             InputStreamReader isr = new InputStreamReader(csvFile.getUploadFile().getInputStream(), StandardCharsets.UTF_8);
@@ -71,87 +68,14 @@ public class FileService {
             return false;
         }
 
-        //clean headers
-        String cleanHeaders = removeSymbols(firstLine);
+        String cleanHeaders = fileUtilService.removeSymbols(firstLine);
         String[] headersFromFile = cleanHeaders.split(",");
-        String[] selectedHeadersArray = this.selectedHeadersArray(csvFile.getSelectedType());
-        boolean isHeadersEquals = isStringArraysEquals(headersFromFile, selectedHeadersArray);
+        String[] selectedHeadersArray = fileUtilService.selectedHeadersArray(csvFile.getSelectedType());
+        boolean isHeadersEquals = fileUtilService.isStringArraysEquals(headersFromFile, selectedHeadersArray);
 
         logger.info("isValidCsvHeader. isHeadersEquals={}", isHeadersEquals);
         return isHeadersEquals;
     }
-
-    public String removeSymbols(String strToClean){
-        String strCleaned = strToClean.replaceAll("[^a-zA-Z0-9,]", "");
-        return strCleaned.toLowerCase();
-    }
-
-    public String[] selectedHeadersArray(String selectedType){
-
-        switch (selectedType){
-
-            case "teacher-roster":
-                return csvHeadersProperties.getTeacherRoster();
-
-            case "teacher-opt-in":
-                return csvHeadersProperties.getTeacherOptIn();
-
-            case "student-level":
-                return csvHeadersProperties.getStudentLevel();
-
-            case "signin-per-course":
-                return csvHeadersProperties.getSigninPerCourse();
-
-            case "sign-ins":
-                return csvHeadersProperties.getSignIns();
-
-            case "diagnostico":
-                return csvHeadersProperties.getDiagnostico();
-
-            case "pre-capacita":
-                return csvHeadersProperties.getPreCapacita();
-
-            case "post-capacita":
-                return csvHeadersProperties.getPostCapacita();
-
-            case "test-pc-1":
-                return csvHeadersProperties.getTestPc1();
-
-            case "test-pc-2":
-                return csvHeadersProperties.getTestPc2();
-
-            case "test-pc-3":
-                return csvHeadersProperties.getTestPc3();
-
-            case "salida":
-                return csvHeadersProperties.getSalida();
-
-            case "satis":
-                return csvHeadersProperties.getSatis();
-
-            default:
-                return new String[1];
-        }
-    }
-
-    public boolean isStringArraysEquals(String[] firstArray, String[] secondArray){
-        logger.info("isStringArraysEquals. firstArray={}", firstArray);
-        logger.info("isStringArraysEquals. secondArray={}", secondArray);
-
-        String[] firstArraySorted = firstArray;
-        String[] secondArraySorted = secondArray;
-
-        Arrays.sort(firstArraySorted);
-        Arrays.sort(secondArraySorted);
-        return Arrays.equals(firstArraySorted, secondArraySorted);
-    }
-
-    public String getFileExtension(String filename){
-        int indexExtensionStart = filename.indexOf(".");
-        String fileExtensionName = filename.substring(indexExtensionStart,filename.length());
-        return fileExtensionName;
-    }
-
 
     public boolean copyAtFileSystem(CsvFile csvFile){
